[Search] Fixed loading only 10 entries, started major cleanup
