+++ lib/modules/search/core.dart
@@ -1,6 +1,7 @@
 export 'core/api.dart';
-export 'core/constants.dart';
 export 'core/database.dart';
 export 'core/dialogs.dart';
+export 'core/models.dart';
+export 'core/router.dart';
 export 'core/sorting.dart';
 export 'core/state.dart';
