--- lib/modules/search/core/api/data.dart
@@ -1,3 +1,0 @@
-export 'data/category.dart';
-export 'data/subcategory.dart';
-export 'data/result.dart';
