--- lib/modules/search/core/constants.dart
@@ -1,5 +1,0 @@
-class SearchConstants {
-    SearchConstants._();
-
-    static const List<String> ADULT_CATEGORIES = ['xxx', 'adult', 'porn'];
-}
