+++ lib/modules/search/core/models.dart
@@ -1,0 +1,3 @@
+export 'models/category.dart';
+export 'models/result.dart';
+export 'models/subcategory.dart';
