+++ lib/modules/search/core/models/result.dart
