+++ lib/modules/search/core/models/subcategory.dart
