+++ lib/modules/search/routes/categories/widgets.dart
@@ -1,0 +1,1 @@
+export 'widgets/category_tile.dart';
