+++ lib/modules/search/routes/categories/widgets/category_tile.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+import 'package:lunasea/core.dart';
+import 'package:lunasea/modules/search.dart';
+
+class SearchCategoryTile extends StatelessWidget {
+    final NewznabCategoryData category;
+    final int index;
+
+    SearchCategoryTile({
+        @required this.category,
+        this.index = 0,
+    });
+
+    @override
+    Widget build(BuildContext context) {
+        return LunaListTile(
+            context: context,
+            title: LunaText.title(text: category.name),
+            subtitle: LunaText.subtitle(
+                text: category.subcategoriesTitleList,
+            ),
+            trailing: LunaIconButton(icon: category.icon, color: LunaColours.list(index)),
+            onTap: () async {
+                context.read<SearchState>().activeCategory = category;
+                context.read<SearchState>().searchCategoryID = category.id;
+                SearchSubcategoriesRouter().navigateTo(context);
+            },
+        );
+    }
+}
