+++ lib/modules/search/routes/indexers/widgets.dart
@@ -1,0 +1,1 @@
+export 'widgets/indexer_tile.dart';
