+++ lib/modules/search/routes/subcategories/widgets.dart
@@ -1,0 +1,2 @@
+export 'widgets/all_subcategories_tile.dart';
+export 'widgets/subcategory_tile.dart';
