+++ lib/modules/search/widgets.dart
@@ -1,9 +1,6 @@
 
 export 'widgets/download_button.dart';
-export 'widgets/category_tile.dart';
 export 'widgets/result_tile.dart';
-export 'widgets/indexer_tile.dart';
 export 'widgets/search_bar.dart';
-export 'widgets/subcategory_tile.dart';
 export 'widgets/result_search_bar.dart';
 export 'widgets/result_sorting_button.dart';
