+++ localization/search/en.json
@@ -1,0 +1,6 @@
+{
+    "search.Categories": "Categories",
+    "search.NoCategoriesFound": "No Categories Found",
+    "search.Search": "Search",
+    "search.Subcategories": "Subcategories"
+}
